+++ tests/tracer/test_monkey.py
@@ -1,3 +1,4 @@
+from ddtrace import _monkey
 from ddtrace import monkey
 from tests.subprocesstest import SubprocessTestCase
 from tests.subprocesstest import run_in_subprocess
@@ -18,18 +19,18 @@ class TestPatching(SubprocessTestCase):
     def test_patch_all_env_override_sqlite_none(self):
         # Make sure sqlite is enabled by default.
         monkey.patch_all()
-        assert "sqlite3" in monkey._PATCHED_MODULES
+        assert "sqlite3" in _monkey._PATCHED_MODULES
 
     @run_in_subprocess(env_overrides=dict(DD_TRACE_SQLITE3_ENABLED="false"))
     def test_patch_all_env_override_sqlite_disabled(self):
         monkey.patch_all()
-        assert "sqlite3" not in monkey._PATCHED_MODULES
+        assert "sqlite3" not in _monkey._PATCHED_MODULES
 
     @run_in_subprocess(env_overrides=dict(DD_TRACE_SQLITE3_ENABLED="false"))
     def test_patch_all_env_override_manual_patch(self):
         # Manual patching should not be affected by the environment variable override.
         monkey.patch(sqlite3=True)
-        assert "sqlite3" in monkey._PATCHED_MODULES
+        assert "sqlite3" in _monkey._PATCHED_MODULES
 
     @run_in_subprocess()
     def test_patch_raise_exception_manual_patch(self):
@@ -41,15 +42,15 @@ class TestPatching(SubprocessTestCase):
             "integration module ddtrace.contrib.module_dne does not exist, module will not have tracing available"
             in str(me.exception)
         )
-        assert "module_dne" not in monkey._PATCHED_MODULES
+        assert "module_dne" not in _monkey._PATCHED_MODULES
 
     @run_in_subprocess(env_overrides=dict())
     def test_patch_all_env_override_httplib_none(self):
         # Make sure httplib is disabled by default.
         monkey.patch_all()
-        assert "httplib" not in monkey._PATCHED_MODULES
+        assert "httplib" not in _monkey._PATCHED_MODULES
 
     @run_in_subprocess(env_overrides=dict(DD_TRACE_HTTPLIB_ENABLED="true"))
     def test_patch_all_env_override_httplib_enabled(self):
         monkey.patch_all()
-        assert "httplib" in monkey._PATCHED_MODULES
+        assert "httplib" in _monkey._PATCHED_MODULES
